Moved user management to an external module
